add other facets

+++ observation.go
@@ -40,9 +40,9 @@ type ObservationResponse struct {
 			QualityControl string  `json:"qualityControl"`
 		} `json:"dewpoint"`
 		WindDirection struct {
-			Value          int    `json:"value"`
-			UnitCode       string `json:"unitCode"`
-			QualityControl string `json:"qualityControl"`
+			Value          float64 `json:"value"`
+			UnitCode       string  `json:"unitCode"`
+			QualityControl string  `json:"qualityControl"`
 		} `json:"windDirection"`
 		WindSpeed struct {
 			Value          float64 `json:"value"`
@@ -55,9 +55,9 @@ type ObservationResponse struct {
 			QualityControl string      `json:"qualityControl"`
 		} `json:"windGust"`
 		BarometricPressure struct {
-			Value          int    `json:"value"`
-			UnitCode       string `json:"unitCode"`
-			QualityControl string `json:"qualityControl"`
+			Value          float64 `json:"value"`
+			UnitCode       string  `json:"unitCode"`
+			QualityControl string  `json:"qualityControl"`
 		} `json:"barometricPressure"`
 		SeaLevelPressure struct {
 			Value          interface{} `json:"value"`
@@ -65,9 +65,9 @@ type ObservationResponse struct {
 			QualityControl string      `json:"qualityControl"`
 		} `json:"seaLevelPressure"`
 		Visibility struct {
-			Value          int    `json:"value"`
-			UnitCode       string `json:"unitCode"`
-			QualityControl string `json:"qualityControl"`
+			Value          float64 `json:"value"`
+			UnitCode       string  `json:"unitCode"`
+			QualityControl string  `json:"qualityControl"`
 		} `json:"visibility"`
 		MaxTemperatureLast24Hours struct {
 			Value          interface{} `json:"value"`
